refactor ManageIssue
