+++ src/__mocks__/BackendAPI.mock.ts
@@ -1,5 +1,4 @@
 import {
-  CompanyId,
   ContributorVisibility,
   FinancialIssue,
   IssueFunding,
