+++ src/components/issue/index.ts
@@ -2,3 +2,4 @@ export * from "./Repository";
 export * from "./Issue";
 export * from "./Collect";
 export * from "./IssueCard";
+export * from "./PrivatePublicToggled";
